requirements.txt
aiohttp==3.12.4
PyYAML==6.0.2
python-dotenv==1.1.0
Flask==3.1.1
flask[async]
Flask-Compress==1.17
